[refactor🔨] : 더미 구조변경에 따른 경로변경, 필요없는 자원제거
